Update and rename Maths/Euclidean Algorithm/Java/factorial.java to Maths/Factorial/Java/factorial.java

+++ Maths/Factorial/Java/factorial.java
@@ -1,12 +1,11 @@
-
 public class fact{
-	static int factorial(int n) {
+	static long factorial(int n) {
 		if(n==0)	return 1;
 		return (n*factorial(n-1));
 	}
 
 	public static void main(String args[]) {
-		int a;
+		long a;
 		a = factorial(4);
 		System.out.println(a);
 	}
